Foxed segfaults and applied organization
