+++ README.md
@@ -1,1 +1,7 @@
 # FingerPassUR Internal Server
+
+## Build
+Execute `go build ./cmd/api` to build the `api`
+
+## Run
+Execute `go run ./cmd/api` to run the `api`
