+++ internal/dto/server_dto.go
@@ -1,0 +1,5 @@
+package dto
+
+type Response struct {
+	Message string `json:"message"`
+}
