+++ internal/dto/user_dto.go
@@ -1,0 +1,8 @@
+package dto
+
+type UserRequest struct {
+	Template string `json:"template"` // Base64 encoded template
+	Correo   string `json:"correo"`
+	Nombre   string `json:"nombre"`
+	Programa string `json:"programa"`
+}
