+++ internal/model/user.go
@@ -1,0 +1,8 @@
+package model
+
+type User struct {
+	Id       int    `json:"id"`
+	Correo   string `json:"correo"`
+	Nombre   string `json:"nombre"`
+	Programa string `json:"programa"`
+}
